refactor:改用DTO 接收前端入参

+++ backend/src/main/java/com/arii/JobTracker/Service/FileServiceImpl.java
@@ -2,6 +2,8 @@ package com.arii.JobTracker.Service;
 
 
 import com.arii.JobTracker.Repository.JobFileRepository;
+import com.arii.JobTracker.Repository.JobRepository;
+import com.arii.JobTracker.Security.SecurityUtils;
 import com.arii.JobTracker.pojo.Job;
 import com.arii.JobTracker.pojo.JobFile;
 import jakarta.transaction.Transactional;
@@ -23,7 +25,10 @@ import java.util.UUID;
 public class FileServiceImpl implements FileService {
     @Autowired
     private JobFileRepository fileRepository;
-
+    @Autowired
+    private SecurityUtils securityUtils;
+    @Autowired
+    private JobRepository jobRepository;
     @Value("${file.upload-path}") // learn;从配置文件读取存放路径
     private String uploadPath;
 
@@ -53,6 +58,16 @@ public class FileServiceImpl implements FileService {
     }
 
     public List<JobFile> getFilesByJobId(Integer jobId) {
+        Integer currentUserId = securityUtils.getCurrentUserId();
+
+        // learn;1. 先查出这个 Job
+        Job job = jobRepository.findById(jobId)
+                .orElseThrow(() -> new RuntimeException("任务不存在"));
+
+        // learn;2. 校验这个 Job 是不是当前用户的
+        if (!job.getUserId().equals(currentUserId)) {
+            throw new RuntimeException("你没有权限查看该任务的附件");
+        }
         return fileRepository.findByJobId(jobId);
     }
 
@@ -63,19 +78,22 @@ public class FileServiceImpl implements FileService {
         JobFile jobFile = fileRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("该文件在数据库中不存在，无法删除"));
 
-        // 2. 拼接该文件在磁盘上的绝对路径 (使用的是长UUID文件名)
+        Integer currentUserId = securityUtils.getCurrentUserId();
+        if (!jobFile.getJob().getUserId().equals(currentUserId)) {
+            throw new RuntimeException("对不起，你没有权限删除该文件！"); // 会触发事务回滚
+        }
+
+        // learn;2. 拼接该文件在磁盘上的绝对路径 (使用的是长UUID文件名)
         java.nio.file.Path filePath = java.nio.file.Paths.get(uploadPath, jobFile.getSavedFileName());
 
         try {
-            // 3. 执行物理删除
             java.nio.file.Files.deleteIfExists(filePath);
             System.out.println("磁盘物理文件删除成功：" + filePath.toString());
         } catch (java.io.IOException e) {
-            // 物理删除失败时抛出异常，触发事务回滚，防止出现“磁盘没删掉，数据库却删了”的情况
+            //learn; 物理删除失败时抛出异常，触发事务回滚，防止出现“磁盘没删掉，数据库却删了”的情况
             throw new RuntimeException("物理文件删除失败，停止清理数据库数据: " + e.getMessage());
         }
 
-        // 4. 执行数据库逻辑删除
         fileRepository.delete(jobFile);
         System.out.println("数据库记录删除成功，文件ID: " + id);
     }
